Implement drag and drop with touch controls
PointerEvent wasn't working - the touchmove event was not properly going through the dragged hex, even with pointer-events: none, so instead, if the pointer moves after dragging starts but the drop target is not specified, keep track of the current mouse position. When the hex is dropped, if the drop target is still not known, find whether there is a hex under the mouse's position and then call handleDrop again, simulating on-grid or off-grid as appropriate.