Fixes #1. Move "incorrect" animation to .hex class and set puzzle to incomplete on drag start
Moving the animation to .hex solved the scaling issue but created a problem where the hexes would wiggle when dragged - probably some style was getting recalculated, but making the containers keyed didn't help. Instead, set the puzzle to Incomplete on drag start so that the animation no longer applies.

Also add Html.Lazy to viewHex because they basically never change.