Change getContiguousHexes and placeHexes to allow from multi-dragging from anywhere
getContiguousHexes now checks all neighborPoints of all hexes in the group, and adds any hexes at those points, but makes sure not to add hexes that have already been found. Because of changes to placeHexes and because not all hexes have an axial position, the axial offset is no longer needed.

placeHexes now calculates the point where the center of each hex is dropped based on the offset from the initial mouse click position and tries to get the axial coordinates of a cell at that point. The canPlace logic is the same with the addition of a check that there are not two hexes fighting for the same spot.

Reversing the list of dragged hexes in the view function prevents z-index fighting when picking up and dropping the same group of hexes.